Major refactoring: Simplify codebase and enhance cardio support
🚀 Major Refactoring Summary:

1. **Enhanced Workout Creation & Encoding**
   - Added comprehensive cardio support with new set types (CARDIO_INTERVAL, CARDIO_STEADY, CARDIO_HIIT)
   - Extended WorkoutSet interface with cardio fields: targetCalories, targetHeartRate, targetPace, targetIncline, targetResistance
   - Enhanced PerformedSet interface with corresponding actual performance fields
   - Improved volume calculation to handle both strength (reps × weight) and cardio (calories or distance × duration)

2. **Consolidated Session Tracking**
   - Removed duplicate API routes: deleted legacy /api/session/route.ts
   - Renamed /api/session-json/route.ts to /api/session/route.ts for unified session management
   - All session data now stored server-side in userStats.activeWorkoutData
   - Removed localStorage migration code and legacy compatibility layers

3. **Removed Legacy/Unused Code**
   - Deleted completely commented out WorkoutForm.tsx (~73 lines)
   - Replaced LegacyPersonalRecord → PersonalRecord and LegacyVolumeRecord → VolumeRecord
   - Cleaned up backward compatibility code from useActiveWorkout.ts
   - Updated workoutDisplayUtils.ts to remove legacy formatting functions

4. **Enhanced Cardio Support**
   - Smart form validation in JsonTemplateForm.tsx based on workout type
   - Cardio sets require duration, distance, or calories; strength sets require reps
   - Enhanced API validation schemas with flexible set validation
   - Proper type checking for cardio-specific fields

5. **Technical Improvements**
   - Fixed TypeScript compilation issues with proper JSON type casting
   - Updated all references to use new type names
   - Maintained backward compatibility for existing workout templates
   - All tests passing with no breaking changes

✅ Benefits: Removed ~500+ lines of legacy code, enhanced cardio functionality, unified session management, improved type safety, and created a cleaner, more maintainable codebase.
