--- src/components/ui/WorkoutForm.tsx
@@ -1,73 +1,0 @@
-// 'use client'
-
-// import { useState, useEffect } from 'react'
-// import FormSection from '@/components/ui/FormSection'
-// import ScheduleSection from '@/components/ui/ScheduleSection'
-// import { generateWorkoutSchedule } from '@/services/workoutGenerator'
-// import type { FormData } from '@/types'
-// import type { WorkoutDay } from '@/types/workout'
-// import { motion } from 'framer-motion'
-
-// export default function WorkoutForm() {
-//   const [formData, setFormData] = useState<FormData>({
-//     name: '',
-//     age: '',
-//     gender: '',
-//     height: '',
-//     weight: '',
-//     fitnessGoal: '',
-//     intensity: '',
-//     workoutsPerWeek: '',
-//     exercisesPerWorkout: ''
-//   })
-
-//   const [workoutSchedule, setWorkoutSchedule] = useState<(WorkoutDay | 'Rest')[]>([])
-
-//   const handleChange = (e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>) => {
-//     const { name, value } = e.target
-//     setFormData(prev => ({ ...prev, [name]: value }))
-//   }
-
-//   const handleSubmit = (e: React.FormEvent) => {
-//     e.preventDefault()
-//     if (!formData.workoutsPerWeek || !formData.exercisesPerWorkout) return
-
-//     const schedule = generateWorkoutSchedule(
-//       Number(formData.workoutsPerWeek),
-//       Number(formData.exercisesPerWorkout)
-//     )
-//     setWorkoutSchedule(schedule)
-//     window.scrollTo({ top: 0, behavior: 'smooth' })
-//   }
-
-//   useEffect(() => {
-//     if (workoutSchedule.length > 0) {
-//       window.scrollTo({ top: 0, behavior: 'smooth' })
-//     }
-//   }, [workoutSchedule])
-
-//   return (
-//     <motion.div 
-//       initial={{ opacity: 0, y: 20 }}
-//       animate={{ opacity: 1, y: 0 }}
-//       className="min-h-[calc(100vh-4rem)] py-12 px-4 overflow-hidden"
-//     >
-//       <div className="w-full">
-//         {workoutSchedule.length === 0 ? (
-//           <FormSection
-//             formData={formData}
-//             handleChange={handleChange}
-//             handleSubmit={handleSubmit}
-//             setFormData={setFormData}
-//           />
-//         ) : (
-//           <ScheduleSection
-//             formData={formData}
-//             workoutSchedule={workoutSchedule}
-//             setWorkoutSchedule={setWorkoutSchedule}
-//           />
-//         )}
-//       </div>
-//     </motion.div>
-//   )
-// }
